Move deferred function calling to rxtx_common
